Move get_formatter from charts/shared.ts to core/format.ts, mark it as legacy with a cleanup/deprecation TODO. Delete the now empty charts/shared.ts
