+++ client/src/panels/panel_declarations/finances/goco/goco.js
@@ -15,10 +15,10 @@ import {
 import { ProgramTag } from "src/models/subjects";
 
 import { newIBCategoryColors } from "src/core/color_schemes";
+import { get_formatter } from "src/core/format";
 import { is_a11y_mode } from "src/core/injected_build_constants";
 
 import { StandardLegend } from "src/charts/legends/index";
-import { get_formatter } from "src/charts/shared";
 import { WrappedNivoBar } from "src/charts/wrapped_nivo/index";
 import { textColor, backgroundColor } from "src/style_constants/index";
 import { Table } from "src/tables/TableClass";
